extract spec_registry
